requirements.txt
async-timeout==4.0.3
bleak==0.21.1
contourpy==1.1.1
cycler==0.12.1
find-libpython==0.3.1
fonttools==4.43.1
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.24.4
packaging==23.2
Pillow==10.1.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-speech-features==0.6
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
zipp==3.17.0
